api/requirements.txt
fastapi
uvicorn[standard]
pydantic
kafka-python
python-dotenv
tenacity